+++ CHANGELOG.md
@@ -4,4 +4,6 @@
 * `/` return the app name, description and version
 * `/version` return the version of the web service
 * User authentication with [stormpath](http://stormpath.com)
-* `/poll/create?choices=1,2,3` Create a new poll (authenticated users only) 
+* `/poll/create?choices={choices}` Create a new poll (authenticated users only)
+* `/poll/createTokens/{pollID}?count={count}` Create new tokens for a poll (Authenticated user admin of the poll only)
+* `/poll/vote/{tokenID}/{secret}?ballot={ballot}` Submit a ballot and consume the vote-token